Implement message by protocol
