+++ main.c
@@ -3,25 +3,34 @@
 #include <mqueue.h>
 #include <sys/stat.h>
 #include <pthread.h>
+#include <string.h>
 #include "chat.h"
 #include "constants.h"
 #include "types.h"
 
-int main(){
+int main(int argc, char *argv[]){
     
     msg_params_t * msg;
     pthread_t receive, send;
     void * thread_res;
     
     msg = (msg_params_t*)malloc(sizeof(msg_params_t));
-    char queue_name[10];
+    char msg_received[500];
+
+    
+    struct mq_attr attr;
+    attr.mq_maxmsg = 10;
+    attr.mq_msgsize = strlen(msg_received);
+    attr.mq_flags = 0;
+ 
     char destination[10];
-    /*
-    if(q_receive = mq_open("/josue",O_RDWR|O_CREAT, 0666, NULL)<0){
-        perror("mq_open");
-        exit(1);
+    char * queue_name = argv[1];
+    if ((q_receive = mq_open (queue_name, O_RDWR|O_CREAT, 0666 , &attr)) < 0){
+        perror ("mq_open");
+        printf("q_receive\n"); 
+        exit (1);
     }
-    */
+
     pthread_create(&send, NULL, send_msg, NULL);
     
     pthread_create(&receive, NULL, receive_msg, (void*) msg);
